requirements.txt
matplotlib
scipy
torch
torchvision
diffusers
pillow
simplejsondb
opencv-python
scikit-image
transformers
accelerate